user login #1

+++ mysite02/src/main/java/mysite/vo/UserVo.java
@@ -1,17 +1,17 @@
 package mysite.vo;
 
 public class UserVo {
-	private int id;
+	private Long id;
 	private String name;
 	private String email;
 	private String password;
 	private String gender;
 	private String joinDate;
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
